# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: Authentication, Company, Job, Statistics, and User API
  description: API for managing user authentication, company records, job postings, statistical data, and user interactions.
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/users/{user_id}:
    get:
      tags:
        - Users
      summary: "Get User Details"
      description: "Retrieve details of a specific user by their ID."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
          description: "The ID of the user to retrieve."
      responses:
        200:
          description: "User details retrieved successfully."
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  email:
                    type: string
                  role:
                    type: string
                    enum: ["admin", "employer", "applicant"]
                  company:
                    type: integer
                    nullable: true
                  created_at:
                    type: string
                    format: date-time
        403:
          description: "Permission denied."
        404:
          description: "User not found."
        500:
          description: "Internal server error."
    put:
      tags:
        - Users
      summary: "Update User Information"
      description: >
        Allows a user to update their own information.
        Admin users can update any user's information, including roles.
        Non-admin users can only update their own information, such as passwords.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
          description: "The ID of the user to update."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: >
                    New password for the user. Must be at least 8 characters long.
                role:
                  type: string
                  enum: ["admin", "employer", "applicant"]
                  description: >
                    New role for the user. Only admins can modify this field.
      responses:
        200:
          description: "User updated successfully."
        400:
          description: >
            Validation error. For example, password does not meet the minimum length.
        403:
          description: >
            Permission denied. Non-admin users cannot update roles or other users' information.
        500:
          description: "Internal server error."
    delete:
      tags:
        - Users
      summary: "Delete User"
      description: "Allows a user to delete their own account. Admins can delete any user's account."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
          description: "The ID of the user to delete."
      responses:
        200:
          description: "User account deleted successfully."
        403:
          description: "Permission denied."
        404:
          description: "User not found."
        500:
          description: "Internal server error."
  /api/users/{user_id}/bookmarks:
    post:
      tags:
        - User Bookmarks
      summary: "Toggle Bookmark"
      description: "Adds or removes a bookmark for a job posting."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
          description: "The ID of the user."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                job_id:
                  type: integer
                  description: "The ID of the job to bookmark."
      responses:
        200:
          description: "Bookmark toggled successfully."
        400:
          description: "Validation error."
        403:
          description: "Permission denied."
    get:
      tags:
        - User Bookmarks
      summary: "List Bookmarks"
      description: "Retrieves the list of bookmarked job postings for a user."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
          description: "The ID of the user."
      responses:
        200:
          description: "Bookmarks retrieved successfully."
        403:
          description: "Permission denied."
  /api/users/{user_id}/applications:
    post:
      tags:
        - User Applications
      summary: "Add Application"
      description: "Adds a new job application for a user."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
          description: "The ID of the user."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                job_id:
                  type: integer
                  description: "The ID of the job to apply for."
                content:
                  type: string
                  description: "The content of the application."
      responses:
        201:
          description: "Application added successfully."
        400:
          description: "Validation error."
        403:
          description: "Permission denied."
    get:
      tags:
        - User Applications
      summary: "List Applications"
      description: "Retrieves the list of job applications for a user."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
          description: "The ID of the user."
      responses:
        200:
          description: "Applications retrieved successfully."
        403:
          description: "Permission denied."
  /api/users/{user_id}/applications/{job_id}:
    delete:
      tags:
        - User Applications
      summary: "Delete Application"
      description: "Deletes a job application for a user."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
          description: "The ID of the user."
        - in: path
          name: job_id
          required: true
          schema:
            type: integer
          description: "The ID of the job application to delete."
      responses:
        200:
          description: "Application deleted successfully."
        403:
          description: "Permission denied."